[level 1] Title: x만큼 간격이 있는 n개의 숫자, Time: 0.04 ms, Memory: 62.3 MB -BaekjoonHub

+++ 프로그래머스/1/12954. x만큼 간격이 있는 n개의 숫자/x만큼 간격이 있는 n개의 숫자.kt
@@ -1,8 +1,8 @@
 class Solution {
-    fun solution(x: Int, n: Int): DoubleArray {
-        var answer = DoubleArray(n)
+    fun solution(x: Int, n: Int): LongArray {
+        var answer = LongArray(n)
         for (i in 1..n){
-            answer[i-1] = x.toDouble() * i.toDouble()
+            answer[i-1] = x.toLong() * i.toLong()
         }
         return answer
     }
